requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# dehazer.py: 35
ImageHash ~= 4.1.0

# dehazer.py: 13
Pillow ~= 4.1.1

# AtmLight.py: 10
# DarkChannel.py: 10
# Radiance.py: 10
# Refine.py: 15
# Transmission.py: 14
numba ~= 0.49

# AImage.py: 10
# dehazer.py: 31
scikit_image ~= 0.15

# AImage.py: 9
# AtmLight.py: 9
# DarkChannel.py: 9
# Dehaze.py: 20
# Radiance.py: 9
# Refine.py: 9
# Transmission.py: 9
# dehazer.py: 11
numpy ~= 1.17
